droid: rename JNI files to JNIManager
