+++ xbmc/android/activity/JNIManager.h
