+++ xbmc/android/activity/Makefile.in
@@ -16,7 +16,7 @@ SRCS      += AndroidMouse.cpp
 SRCS      += GraphicBuffer.cpp
 SRCS      += EventLoop.cpp
 SRCS      += XBMCApp.cpp
-SRCS      += JNI.cpp
+SRCS      += JNIManager.cpp
 SRCS      += Intents.cpp
 SRCS      += JNIThreading.cpp
 SRCS      += BroadcastReceiver.cpp
